requirements.txt
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
fastai==2.4
fastapi==0.65.1
fastcore==1.3.20
fastprogress==1.0.0
h11==0.12.0
idna==2.10
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
murmurhash==1.0.5
numpy==1.21.0
packaging==21.0
pandas==1.3.0
pathy==0.6.0
Pillow==8.2.0
preshed==3.0.5
pydantic==1.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
smart-open==5.1.0
spacy==3.0.6
spacy-legacy==3.0.6
srsly==2.4.1
starlette==0.14.2
thinc==8.0.7
threadpoolctl==2.1.0
torch
torchvision==0.10.0
tqdm==4.61.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.13.4
wasabi==0.8.2
